+++ core/src/error.rs
@@ -96,11 +96,13 @@ impl From<casper_types::bytesrepr::Error> for ExecutionError {
 ///
 /// It is represented by an error code and a human-readable message.
 ///
-/// Errors codes 0..32767 are available for the user to define custom error
+/// Errors codes 0..64534 are available for the user to define custom error
 /// in smart contracts.
-/// 32768 code is a special code representing a violation of the custom error code space.
+/// 64535 code (`MaxUserError`) is the upper bound of the custom error code space and
+/// 64536 (`UserErrorTooHigh`) is reported when a custom error exceeds it.
 ///
-/// The rest of codes 32769..[u16::MAX](u16::MAX), are used internally by the framework.
+/// The rest of codes 64537..[u16::MAX](u16::MAX) are used internally by the framework.
+/// The on-chain code of an internal error is `64536 + discriminant`.
 #[repr(u16)]
 #[derive(Clone, Debug, PartialEq)]
 pub enum ExecutionError {